requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
web3>=6.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
shap>=0.43.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0
tqdm>=4.65.0
plotly>=5.14.0
